solutions/sbom-chatbot/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
astroid==3.2.4
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.2
boto3==1.34.160
botocore==1.34.160
cachetools==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
identify==2.6.0
idna==3.7
isort==5.13.2
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
llama-cloud==0.0.13
llama-index==0.10.65
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.65
llama-index-embeddings-bedrock==0.2.1
llama-index-embeddings-openai==0.1.11
llama-index-graph-stores-neptune==0.1.4
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post1
llama-index-llms-anthropic==0.1.17
llama-index-llms-bedrock==0.1.12
llama-index-llms-openai==0.1.29
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.4.1
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.2
nodeenv==1.9.1
numpy==1.26.4
openai==1.40.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
protobuf==5.27.3
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pylint==3.2.6
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
st-cytoscape==0.0.5
streamlit==1.37.1
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wrapt==1.16.0
yarl==1.9.4
